requirements.txt
Flask>=2.3,<4
openai>=1,<2
groq>=0.5,<1
python-dotenv>=1,<2
mysql-connector-python>=8.0,<9